Modified .gitignore for .class files, updated readme, subbed long in QQ.

+++ core/src/com/ridicarus/kid/tools/QQ.java
@@ -63,7 +63,7 @@ public class QQ {
 	}
 
 	// current time millis mod a big number, so we see the time without e notation
-	public static int hackTime() {
+	public static long hackTime() {
 		return Math.floorMod(System.currentTimeMillis(), 100000);
 	}
 }
